requirements.txt
## apache-airflow==2.9.3
apache-airflow-providers-http
apache-airflow-providers-postgres
psycopg2-binary
